ver log, priority sort,priority tasks

+++ projects.js
@@ -840,16 +840,16 @@ function enableColumnSorting() {
 
                 // Gestione speciale per la colonna priority (indice 12)
                 if (columnIndex === 12) {
-                    // Converti in numeri, se non è un numero sarà NaN
-                    const aNum = parseInt(aText);
-                    const bNum = parseInt(bText);
+                    // Converti in numeri se possibile
+                    const aNum = parseFloat(aText);
+                    const bNum = parseFloat(bText);
                     
                     // Se entrambi sono numeri, confronta numericamente
                     if (!isNaN(aNum) && !isNaN(bNum)) {
                         return isAscending ? aNum - bNum : bNum - aNum;
                     }
                     
-                    // Se solo uno è numero, quello va prima
+                    // Se solo uno è numero, il numero va prima
                     if (!isNaN(aNum)) return isAscending ? -1 : 1;
                     if (!isNaN(bNum)) return isAscending ? 1 : -1;
                     
@@ -857,7 +857,7 @@ function enableColumnSorting() {
                     return isAscending ? aText.localeCompare(bText) : bText.localeCompare(aText);
                 }
 
-                // Per tutte le altre colonne, usa il confronto normale
+                // Per tutte le altre colonne, mantieni il comportamento originale
                 return isAscending ? aText.localeCompare(bText) : bText.localeCompare(aText);
             });
